+++ amazon-chime-sdk/src/test/java/com/amazonaws/services/chime/sdk/meetings/internal/audio/DefaultAudioClientObserverTest.kt
@@ -17,6 +17,7 @@ import com.amazonaws.services.chime.sdk.meetings.audiovideo.SignalStrength
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.SignalUpdate
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.Transcript
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.TranscriptAlternative
+import com.amazonaws.services.chime.sdk.meetings.audiovideo.TranscriptEntity
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.TranscriptItem
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.TranscriptItemType
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.TranscriptResult
@@ -37,6 +38,7 @@ import com.xodee.client.audio.audioclient.AttendeeUpdate
 import com.xodee.client.audio.audioclient.AudioClient
 import com.xodee.client.audio.audioclient.transcript.Transcript as TranscriptInternal
 import com.xodee.client.audio.audioclient.transcript.TranscriptAlternative as TranscriptAlternativeInternal
+import com.xodee.client.audio.audioclient.transcript.TranscriptEntity as TranscriptEntityInternal
 import com.xodee.client.audio.audioclient.transcript.TranscriptEvent as TranscriptEventInternal
 import com.xodee.client.audio.audioclient.transcript.TranscriptItem as TranscriptItemInternal
 import com.xodee.client.audio.audioclient.transcript.TranscriptItemType as TranscriptItemTypeInternal
@@ -644,7 +646,9 @@ class DefaultAudioClientObserverTest {
                             timestampMs + 5L,
                             AttendeeInfoInternal(testId1, testId1),
                             "I",
-                            true
+                            true,
+                            true,
+                            0.0
                         ),
                         TranscriptItemInternal(
                             TranscriptItemTypeInternal.TranscriptItemTypePunctuation,
@@ -652,9 +656,12 @@ class DefaultAudioClientObserverTest {
                             timestampMs + 10L,
                             AttendeeInfoInternal(testId2, testId2),
                             "am",
-                            false
+                            false,
+                            true,
+                            0.0
                         )
                     ),
+                    arrayOf(),
                     "I am"
                 )
             )
@@ -675,7 +682,9 @@ class DefaultAudioClientObserverTest {
                             timestampMs + 15L,
                             AttendeeInfoInternal(testId2, testId2),
                             "a",
-                            true
+                            true,
+                            true,
+                            0.0
                         ),
                         TranscriptItemInternal(
                             TranscriptItemTypeInternal.TranscriptItemTypePronunciation,
@@ -683,9 +692,12 @@ class DefaultAudioClientObserverTest {
                             timestampMs + 20L,
                             AttendeeInfoInternal(testId1, testId1),
                             "guardian",
-                            false
+                            false,
+                            true,
+                            0.0
                         )
                     ),
+                    arrayOf(),
                     "a guardian"
                 )
             )
@@ -713,6 +725,8 @@ class DefaultAudioClientObserverTest {
                                     timestampMs + 5L,
                                     AttendeeInfo(testId1, testId1),
                                     "I",
+                                    true,
+                                    0.0,
                                     true
                                 ),
                                 TranscriptItem(
@@ -721,9 +735,12 @@ class DefaultAudioClientObserverTest {
                                     timestampMs + 10L,
                                     AttendeeInfo(testId2, testId2),
                                     "am",
-                                    false
+                                    false,
+                                    0.0,
+                                    true
                                 )
                             ),
+                            arrayOf(),
                             "I am"
                         )
                     )
@@ -748,6 +765,8 @@ class DefaultAudioClientObserverTest {
                                     timestampMs + 15L,
                                     AttendeeInfo(testId2, testId2),
                                     "a",
+                                    true,
+                                    0.0,
                                     true
                                 ),
                                 TranscriptItem(
@@ -756,9 +775,12 @@ class DefaultAudioClientObserverTest {
                                     timestampMs + 20L,
                                     AttendeeInfo(testId1, testId1),
                                     "guardian",
-                                    false
+                                    false,
+                                    0.0,
+                                    true
                                 )
                             ),
+                            arrayOf(),
                             "a guardian"
                         )
                     )
@@ -769,6 +791,40 @@ class DefaultAudioClientObserverTest {
         audioClientObserver.onTranscriptEventsReceived(events)
         verify(exactly = 1) { mockTranscriptEventObserver.onTranscriptEventReceived(expectedTranscriptOne) }
         verify(exactly = 1) { mockTranscriptEventObserver.onTranscriptEventReceived(expectedTranscriptTwo) }
+    }
+
+    @Test
+    fun `onTranscriptEventsReceived should send local Transcript with Transcript Entity events as input`() {
+        audioClientObserver.subscribeToTranscriptEvent(mockTranscriptEventObserver)
+
+        val testResultId = "testResultId"
+        val testChannelId = "testChannelId"
+        val isPartial = true
+
+        val transcriptResultItem = TranscriptItemInternal(TranscriptItemTypeInternal.TranscriptItemTypePronunciation,
+            timestampMs, timestampMs + 5L, AttendeeInfoInternal(testId1, testId1), "I", true, true, 0.0)
+
+        val transcriptResultEntity = TranscriptEntityInternal("PII", "NAME", "John Doe", 1.0, timestampMs + 5L, timestampMs + 10L)
+
+        val transcriptResultAlternative = TranscriptAlternativeInternal(arrayOf(transcriptResultItem), arrayOf(transcriptResultEntity), "I am")
+
+        val transcriptResult = TranscriptResultInternal(testResultId, testChannelId, isPartial, timestampMs, timestampMs + 10L, arrayOf(transcriptResultAlternative))
+
+        val events: Array<TranscriptEventInternal> = arrayOf(TranscriptInternal(arrayOf(transcriptResult)))
+
+        val expectedTranscriptItem = TranscriptItem(TranscriptItemType.Pronunciation, timestampMs, timestampMs + 5L,
+            AttendeeInfo(testId1, testId1), "I", true, 0.0, true)
+
+        val expectedTranscriptEntity = TranscriptEntity("NAME", 1.0, "John Doe", timestampMs + 5L, timestampMs + 10L, "PII")
+
+        val expectedTranscriptAlternative = TranscriptAlternative(arrayOf(expectedTranscriptItem), arrayOf(expectedTranscriptEntity),
+            "I am")
+
+        val expectedTranscript = Transcript(arrayOf(TranscriptResult(testResultId, testChannelId, isPartial,
+            timestampMs, timestampMs + 10L, arrayOf(expectedTranscriptAlternative))))
+
+        audioClientObserver.onTranscriptEventsReceived(events)
+        verify(exactly = 1) { mockTranscriptEventObserver.onTranscriptEventReceived(expectedTranscript) }
     }
 
     @Test