+++ app/src/main/java/com/amazonaws/services/chime/sdkdemo/data/TranscriptionConfig.kt
@@ -1,19 +1,16 @@
 package com.amazonaws.services.chime.sdkdemo.data
 
-data class TranscribeEngine(val engine: String, val name: String) {
+data class SpinnerItem(val spinnerText: String?, val value: String) {
     override fun toString(): String {
-        return name
+        return value
     }
 }
 
-data class TranscribeLanguage(val code: String, val name: String) {
-    override fun toString(): String {
-        return name
-    }
-}
-
-data class TranscribeRegion(val code: String, val name: String) {
-    override fun toString(): String {
-        return name
-    }
-}
+data class TranscriptionStreamParams(
+    val contentIdentificationType: String?,
+    val contentRedactionType: String?,
+    val enablePartialResultsStabilization: Boolean,
+    val partialResultsStability: String?,
+    val piiEntityTypes: String?,
+    val languageModelName: String?
+)